+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Missing Permissions on upgrade from Juniper to Kiwi. (UIEH-1231)
 * Fix datepicker date format for Custom coverage dates to match current user locale when edit resource. (UIEH-1227)
 * Add tests for useMultiColumnListSort. (UIEH-1183)
+* Add tests for ResourceEditRoute. (UIEH-1190)
 * Fix multiple requests to /resources sent when viewing Titles list on Package Show page. (UIEH-1235)
 
 ## [7.0.1] (https://github.com/folio-org/ui-eholdings/tree/v7.0.1) (2021-10-19)